refactor(cli): extract manifest egress domain
