+++ packages/cli/src/runtime/manifest.ts
@@ -1,7 +1,5 @@
-import { createHash } from "node:crypto";
 import {
 	chmodSync,
-	chownSync,
 	existsSync,
 	lstatSync,
 	mkdirSync,
@@ -13,7 +11,6 @@ import {
 import { dirname, isAbsolute, join, relative, resolve } from "node:path";
 import { parse as parseYaml } from "yaml";
 import { buildAgentTargetProjection } from "../lib/ai-provider-projection";
-import { resolveCurrentCliResourceRoot } from "../lib/current-cli-invocation";
 import { writePrivateFileAtomic } from "../lib/private-file";
 import { readRuntimeAppliedState, runtimeContentSha256 } from "./applied-state";
 import {
@@ -110,6 +107,17 @@ import {
 	type RuntimeManifest,
 } from "./manifest-contract";
 import {
+	clearEgressAddon,
+	clearEgressProfileBundle,
+	ensureRuntimeUserCliStateRoot,
+	makeEgressIdentityPrivateDir,
+	requireV2EgressEngineReady,
+	writeEgressAddon,
+	writeEgressEngineStatus,
+	writeEgressProfileBundle,
+	writeTransparentEgressEnvFile,
+} from "./manifest-egress";
+import {
 	commitRuntimeInstallReceipts,
 	observeRuntimeInstall,
 	planRuntimeInstallObservation,
@@ -161,7 +169,6 @@ import {
 } from "./manifest-runtime-config";
 import {
 	egressSecretFilePath,
-	makeEgressIdentityOwned,
 	makeManagedSecretRoot,
 	type RuntimeEgressSecretMaterial,
 	scopedSecretValues,
@@ -195,18 +202,10 @@ export {
 	type RuntimeManifestLoad,
 } from "./manifest-source";
 
-import {
-	buildEgressProfileBundle,
-	type EgressProfileBundle,
-	hasEnabledEgressProfiles,
-} from "./egress-profiles";
+import { buildEgressProfileBundle, hasEnabledEgressProfiles } from "./egress-profiles";
 import type { RuntimeManifestLoad } from "./manifest-source";
-import {
-	type EnsureRuntimeMitmproxyOptions,
-	ensureRuntimeMitmproxy,
-	type RuntimeMitmproxyEnsureResult,
-} from "./mitmproxy-fetch";
-import { RUNTIME_USER_CLI_STATE_ROOT_MODE, type RuntimePaths } from "./paths";
+import { type EnsureRuntimeMitmproxyOptions, ensureRuntimeMitmproxy } from "./mitmproxy-fetch";
+import type { RuntimePaths } from "./paths";
 import { hostedRuntimeProjectionHome } from "./projection-home";
 import {
 	buildRuntimeRunConfig,
@@ -241,18 +240,11 @@ import {
 	executableExists,
 	makeRuntimeUserOwned,
 	type RuntimeUserOwnershipRule,
-	runningAsRoot,
-	runtimeEgressGid,
 	runtimeUserDirectoryOwnership,
 	runtimeUserExistingOwnership,
 	withRuntimeUserFileAccess,
 } from "./runtime-user-command";
 import { ensureRuntimePlatformDirectory } from "./state";
-
-import {
-	TRANSPARENT_EGRESS_TABLE,
-	TRANSPARENT_EGRESS_TRANSPORT_VERSION,
-} from "./transparent-egress";
 
 export interface RuntimeResourcePreparationFailures {
 	agentPlugins?: {
@@ -284,173 +276,6 @@ export function planHostedAgentPluginConvergence(input: {
 			...(input.runner ? { runner: input.runner } : {}),
 		}),
 	};
-}
-
-function ensureRuntimeUserCliStateRoot(path: string, identity: { uid: number; gid: number }): void {
-	mkdirSync(path, { recursive: true });
-	let node = lstatSync(path);
-	if (!node.isDirectory() || node.isSymbolicLink()) {
-		throw new Error(`hosted CLAWDI_HOME must be a real directory: ${path}`);
-	}
-	if (runningAsRoot()) chownSync(path, identity.uid, identity.gid);
-	chmodSync(path, RUNTIME_USER_CLI_STATE_ROOT_MODE);
-	node = lstatSync(path);
-	if (
-		(node.mode & 0o777) !== RUNTIME_USER_CLI_STATE_ROOT_MODE ||
-		node.uid !== identity.uid ||
-		node.gid !== identity.gid
-	) {
-		throw new Error(`hosted CLAWDI_HOME ownership or mode is invalid: ${path}`);
-	}
-}
-
-function makeEgressIdentityPrivateDir(path: string): void {
-	mkdirSync(path, { recursive: true });
-	makeEgressIdentityOwned(path);
-	try {
-		chmodSync(path, 0o700);
-	} catch {
-		// Best effort for non-POSIX local development environments.
-	}
-}
-
-function clearEgressProfileBundle(paths: RuntimePaths): null {
-	rmSync(paths.egressProfileBundle, { force: true });
-	return null;
-}
-
-function writeEgressEngineStatus(
-	result: RuntimeMitmproxyEnsureResult | null,
-	paths: RuntimePaths,
-): RuntimeMitmproxyEnsureResult | null {
-	if (!result) {
-		rmSync(paths.egressEngineStatus, { force: true });
-		return null;
-	}
-	writeJsonFile(paths.egressEngineStatus, result, paths);
-	return result;
-}
-
-function requireV2EgressEngineReady(
-	manifest: RuntimeManifest,
-	profileBundlePath: string | null,
-	engine: RuntimeMitmproxyEnsureResult | null,
-): void {
-	if (
-		manifest.projection?.sourceBundleVersion === HOSTED_RUNTIME_BUNDLE_V2_SCHEMA_VERSION &&
-		profileBundlePath &&
-		engine?.status !== "ready"
-	) {
-		throw new Error(
-			`required egress engine is not ready: ${engine?.error ?? "status unavailable"}`,
-		);
-	}
-}
-
-function writeEgressProfileBundle(bundle: EgressProfileBundle, paths: RuntimePaths): string {
-	// Published handoff: the egress sidecar (clawdi-egress uid) reads this
-	// bundle, so it lives under the traversable run root next to the other
-	// sidecar inputs (addon, transparent env, CA) — never under a private
-	// platform root the sidecar cannot traverse.
-	writeRuntimePrivateFileAtomic(
-		paths,
-		paths.egressProfileBundle,
-		`${JSON.stringify(bundle, null, 2)}\n`,
-		{
-			mode: 0o640,
-			dirMode: 0o711,
-		},
-	);
-	if (runningAsRoot()) chownSync(paths.egressProfileBundle, 0, runtimeEgressGid());
-	return paths.egressProfileBundle;
-}
-
-function writeEgressAddon(paths: RuntimePaths): { path: string; sha256: string } {
-	const source = resolvePackagedEgressAddon();
-	const content = readFileSync(source, "utf-8");
-	writeRuntimePrivateFileAtomic(paths, paths.egressAddon, content, {
-		mode: 0o640,
-		dirMode: 0o711,
-	});
-	if (runningAsRoot()) chownSync(paths.egressAddon, 0, runtimeEgressGid());
-	return { path: paths.egressAddon, sha256: sha256String(content) };
-}
-
-function clearEgressAddon(paths: RuntimePaths): null {
-	rmSync(paths.egressAddon, { force: true });
-	return null;
-}
-
-function resolvePackagedEgressAddon(): string {
-	const candidate = resolve(
-		resolveCurrentCliResourceRoot(),
-		"egress-addon",
-		"clawdi_egress_addon.py",
-	);
-	if (existsSync(candidate)) return candidate;
-	throw new Error("packaged egress addon is missing");
-}
-
-function writeTransparentEgressEnvFile(input: {
-	program: RuntimeEgressSystemdProgram | null;
-	paths: RuntimePaths;
-	runtimeUser: string;
-	runtimeUid: number;
-	runtimeGid: number;
-	egressUid: number;
-	egressGid: number;
-}): string | null {
-	if (!input.program) {
-		rmSync(input.paths.egressTransparentEnv, { force: true });
-		return null;
-	}
-	const env: Record<string, string> = {
-		CLAWDI_RUNTIME_USER: input.runtimeUser,
-		CLAWDI_RUNTIME_UID: String(input.runtimeUid),
-		CLAWDI_RUNTIME_GID: String(input.runtimeGid),
-		CLAWDI_EGRESS_UID: String(input.egressUid),
-		CLAWDI_EGRESS_GID: String(input.egressGid),
-		CLAWDI_EGRESS_TRANSPARENT_PORT: String(input.program.transparentPort),
-		CLAWDI_EGRESS_NFT_TABLE: TRANSPARENT_EGRESS_TABLE,
-		CLAWDI_EGRESS_PROFILE_BUNDLE: input.program.profileBundlePath,
-		CLAWDI_EGRESS_SECRET_FILE: input.program.secretFilePath ?? "",
-		CLAWDI_EGRESS_CA_DIR: input.paths.egressCaDir,
-		CLAWDI_EGRESS_CA_CERT: input.paths.egressCaCert,
-		CLAWDI_EGRESS_SYSTEM_CA_BUNDLE: input.program.systemCaBundle,
-		CLAWDI_EGRESS_TRANSPORT_VERSION: TRANSPARENT_EGRESS_TRANSPORT_VERSION,
-		CLAWDI_EGRESS_ENGINE_TYPE: "mitmproxy",
-		CLAWDI_EGRESS_ENGINE_VERSION: input.program.engine.version,
-		CLAWDI_EGRESS_ENGINE_URL: input.program.engine.url,
-		CLAWDI_EGRESS_ENGINE_SHA256: input.program.engine.sha256,
-		CLAWDI_EGRESS_ENGINE_BINARY_PATH: input.paths.egressServiceBinary,
-		CLAWDI_EGRESS_ADDON_PATH: input.program.addonPath,
-		CLAWDI_EGRESS_ADDON_SHA256: input.program.addonSha256,
-	};
-	const lines = Object.entries(env)
-		.sort(([a], [b]) => a.localeCompare(b))
-		.map(([key, value]) => `${key}=${runtimeEnvironmentFileQuote(value)}`);
-	writeRuntimePrivateFileAtomic(
-		input.paths,
-		input.paths.egressTransparentEnv,
-		`${lines.join("\n")}\n`,
-		{
-			mode: 0o640,
-			dirMode: 0o711,
-		},
-	);
-	if (runningAsRoot()) chownSync(input.paths.egressTransparentEnv, 0, input.egressGid);
-	return input.paths.egressTransparentEnv;
-}
-
-function runtimeEnvironmentFileQuote(value: string): string {
-	if (/[\r\n]/.test(value)) {
-		throw new Error("runtime environment files only support single-line values");
-	}
-	return `"${value.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}"`;
-}
-
-function sha256String(value: string): string {
-	return createHash("sha256").update(value).digest("hex");
 }
 
 function removeStaleRuntimeRunConfigs(writtenRunConfigIds: Set<string>, paths: RuntimePaths): void {
